requirements.txt
pandas
matplotlib
seaborn
sqlalchemy
fpdf2
